clusterworkspace: quality initializer names
We need globally unique identifiers for the initializing entities, or
our model is insecure. By adding the workspace path from which an
initializer runs to the name of the initializer, we get such an
identifier.

In addition, we change the labelling mechanisms to use a hash of these
values in order to ensure that we continue to create unique and correct
label keys and values.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>